DDBTEAM-640: Moved cover user upload processor into ampq
