+++ config/packages/messenger.yaml
@@ -46,4 +46,5 @@ framework:
             'App\Message\DeleteMessage': background-queue
             'App\Message\VendorImageMessage': cover-store-queue
             'App\Message\CoverStoreMessage': cover-store-queue
+            'App\Message\CoverUserUploadMessage': cover-store-queue
             'App\Message\CoverStoreAutoMessage': cover-store-auto-queue
