+++ src/MessageHandler/CoverUserUploadMessageHandler.php
@@ -5,31 +5,27 @@
  * Upload image service queue processor.
  */
 
-namespace App\Queue;
+namespace App\MessageHandler;
 
 use App\Entity\Source;
 use App\Entity\Vendor;
 use App\Event\VendorEvent;
 use App\Exception\IllegalVendorServiceException;
 use App\Exception\UnknownVendorServiceException;
+use App\Message\CoverUserUploadMessage;
 use App\Repository\SourceRepository;
 use App\Repository\VendorRepository;
 use App\Service\VendorService\UserUpload\UserUploadVendorService;
-use App\Utils\Message\CoverUploadProcessMessage;
 use App\Utils\Types\VendorState;
 use Doctrine\ORM\EntityManagerInterface;
-use Enqueue\Client\TopicSubscriberInterface;
-use Interop\Queue\Context;
-use Interop\Queue\Message;
-use Interop\Queue\Processor;
-use Karriere\JsonDecoder\JsonDecoder;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 
 /**
  * Class CoverUploadProcessor.
  */
-class CoverUserUploadProcessor implements Processor, TopicSubscriberInterface
+class CoverUserUploadMessageHandler implements MessageHandlerInterface
 {
     private $em;
     private $dispatcher;
@@ -64,25 +60,24 @@ class CoverUserUploadProcessor implements Processor, TopicSubscriberInterface
     }
 
     /**
-     * {@inheritdoc}
+     * @param CoverUserUploadMessage $message
+     *
+     * @throws \Doctrine\ORM\Query\QueryException
      */
-    public function process(Message $message, Context $session)
+    public function __invoke(CoverUserUploadMessage $message)
     {
-        $jsonDecoder = new JsonDecoder(true);
-        /** @var CoverUploadProcessMessage $uploadProcessMessage */
-        $uploadProcessMessage = $jsonDecoder->decode($message->getBody(), CoverUploadProcessMessage::class);
-        $identifier = $uploadProcessMessage->getIdentifier();
+        $identifier = $message->getIdentifier();
 
         /** @var Source[] $sources */
-        $sources = $this->sourceRepo->findByMatchIdList($uploadProcessMessage->getIdentifierType(), [$identifier => ''], $this->vendor);
+        $sources = $this->sourceRepo->findByMatchIdList($message->getIdentifierType(), [$identifier => ''], $this->vendor);
 
-        $event = new VendorEvent(VendorState::UNKNOWN, [$identifier], $uploadProcessMessage->getIdentifierType(), $this->vendor->getId());
+        $event = new VendorEvent(VendorState::UNKNOWN, [$identifier], $message->getIdentifierType(), $this->vendor->getId());
 
-        switch ($uploadProcessMessage->getOperation()) {
+        switch ($message->getOperation()) {
             case VendorState::UPDATE:
             case VendorState::INSERT:
                 $event->changeType(VendorState::UPDATE);
-                if ($this->createUpdateSource($identifier, $sources, $uploadProcessMessage)) {
+                if ($this->createUpdateSource($identifier, $sources, $message)) {
                     $event->changeType(VendorState::INSERT);
                 }
                 break;
@@ -92,9 +87,7 @@ class CoverUserUploadProcessor implements Processor, TopicSubscriberInterface
                 break;
         }
 
-        $this->dispatcher->dispatch($event::NAME, $event);
-
-        return self::ACK;
+        $this->dispatcher->dispatch($event, $event::NAME);
     }
 
     /**
@@ -104,13 +97,13 @@ class CoverUserUploadProcessor implements Processor, TopicSubscriberInterface
      *   Material identifier (matchId)
      * @param Source[] $sources
      *   The sources found based on the identifier in the database
-     * @param CoverUploadProcessMessage $uploadProcessMessage
+     * @param CoverUserUploadMessage $uploadProcessMessage
      *   The process message to build for the event producer
      *
      * @return bool
      *   true on insert and false on update
      */
-    private function createUpdateSource(string $identifier, array $sources, CoverUploadProcessMessage $uploadProcessMessage): bool
+    private function createUpdateSource(string $identifier, array $sources, CoverUserUploadMessage $uploadProcessMessage): bool
     {
         $isNew = true;
         if (array_key_exists($identifier, $sources)) {
@@ -140,21 +133,4 @@ class CoverUserUploadProcessor implements Processor, TopicSubscriberInterface
 
         return $isNew;
     }
-
-    // phpcs:disable Symfony.Functions.ScopeOrder.Invalid
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function getSubscribedTopics(): array
-    {
-        return [
-            'UserUploadImageTopic' => [
-                'processorName' => 'CoverUserUploadProcessor',
-                'queueName' => 'CoverStoreQueue',
-            ],
-        ];
-    }
-
-    // phpcs:enable
 }
